Mayor restructure: separate sub-projects for widget, designer plugin. Add two examples. Add install rules for Unix like systems. Tested to build and run with Qt5.4.1 and Qt.5.6.0 on Ubuntu 16.04.
