+++ examples/examples.pro
@@ -1,0 +1,2 @@
+TEMPLATE = subdirs
+SUBDIRS  = simple contextMenu
